[IPS-2492] refactor: Restructure tests
