+++ tests/conftest.py
@@ -6,10 +6,6 @@ from io import FileIO
 from pathlib import Path
 from typing import IO
 
-from ips_client.data_models import ServiceType, OutputType
-from ips_client.ips_instance import IPSInstance
-from ips_client.ips_requests import IPSRequests
-from ips_client.job import IPSJob
 from ips_client.settings import Settings
 
 
@@ -50,23 +46,6 @@ def some_custom_lp_stamp() -> IO:
     img_path = pathlib.Path(__file__).parent.joinpath('resources/licence_plate_custom_stamp.png')
     with open(img_path, 'rb') as f:
         yield f
-
-
-@pytest.fixture
-def ips_requests(ips_url) -> IPSRequests:
-    return IPSRequests(ips_url=ips_url)
-
-
-@pytest.fixture(params=[ServiceType.dnat, ServiceType.blur, ServiceType.extract])
-def any_img_ips(ips_url, request) -> IPSInstance:
-    service = request.param
-    out_type = OutputType.overlays if service == ServiceType.extract else OutputType.images
-    return IPSInstance.create(service=service, out_type=out_type, ips_url=ips_url)
-
-
-@pytest.fixture
-def job(any_img_ips, some_image) -> IPSJob:
-    return any_img_ips.start_job(file=some_image)
 
 
 @pytest.fixture
