+++ tests/functional/test_anonymize_folder.py
