+++ tests/functional/test_custom_stamp.py
